fix(keymaps): solved mapping issue
The issue was the `packer_init.lua` was being sourced before `editor.lua` in `init.lua`. This caused `telescopeconfig.lua` to be sourced before the `leader>` was set in `editor.lua`. The issue was solved by sourcing the files in the correct order
The issue affecting lsp was that diagnostics were being mapped before
on_attach function
Signed-off-by: giluis <lupevibogil@gmail.com>